Add new main logo file
Create a new set of digit shapes, implemented as clones of simple
objects, reducing effort for future iterations of the event. The
digits were not aligned correctly to the negative space. Make
them center aligned.

The SVG viewbox was a little bit wider than it needed to be. Make it
take the minimum required width.

Fixes https://github.com/fossunited/Branding/issues/9